removed code that was not needed
