+++ starwars/src/App.js
@@ -1,7 +1,7 @@
 import React, {useState, useEffect} from 'react';
 import axios from 'axios';
 import './App.css';
-import CharacterCard from './components/CharacterCard'
+import Character from './components/Character'
 import {BASE_URL} from './constants/index'
 import styled from 'styled-components'
 
@@ -9,6 +9,10 @@ const AppStyled = styled.div`
   background-image: url('https://images.unsplash.com/photo-1520034475321-cbe63696469a?ixlib=rb-1.2.1&ixid=eyJhcHBfaWQiOjEyMDd9&auto=format&fit=crop&w=800&q=60');
   margin: 0 3%;
   padding: 3%;
+`
+
+const StyledH1 = styled.h1`
+    color: white;
 `
 
 const App = () => {
@@ -33,9 +37,40 @@ const App = () => {
       return ()=>{}
   }, []);
 
+  const mapCharacterNames = props => {
+    let names = [];
+
+    Object.values(props).map((values, index) => {
+        Object.keys(props[index]).map(key => {
+            if( key === 'name'){
+                names.push(props[index][key]);
+            }
+        })
+    })
+    return names;
+  }
+
+  const mapCharacterBdays = props => {
+      let bdays = [];
+
+      Object.values(props).map((values, index) => {
+          Object.keys(props[index]).map(key => {
+              if( key === 'birth_year'){
+                  bdays.push(props[index][key]);
+              }
+          })
+      })
+      return bdays;
+  }
+
   return (
     <AppStyled className="App">
-      <CharacterCard characterList={characterList}/>
+       <StyledH1>Characters</StyledH1>
+            {
+                mapCharacterNames(characterList).map((name, index) => {
+                    return <Character name={name} bday ={mapCharacterBdays(characterList)[index]}/>
+                })
+            }
     </AppStyled>
   );
 }
